removed helper update.py + moved static stats tests to unit test
